requirements.txt
tensorflow>= 2.0.0.a0
keras==2.3.0
torchvision
numpy
pillow
sklearn
imblearn
# torch-scatter -f https://pytorch-geometric.com/whl/torch-1.9.0+cu102.html 
# torch-sparse -f https://pytorch-geometric.com/whl/torch-1.9.0+cu102.html
# torch-cluster -f https://pytorch-geometric.com/whl/torch-1.9.0+cu102.html
torch-geometric
torchgeometry
dgl
torchmetrics
